organize godot project, add gltf container
