+++ godot/src/main.tscn
@@ -1,252 +1,6 @@
-[gd_scene load_steps=12 format=3 uid="uid://deq5v42fmh0y7"]
+[gd_scene load_steps=2 format=3 uid="uid://b0jonmd2rfupn"]
 
 [ext_resource type="Script" path="res://src/main.gd" id="1_2a1rx"]
-[ext_resource type="Script" path="res://src/camera.gd" id="2_ray12"]
-[ext_resource type="Texture2D" uid="uid://m7i23bc2yyqt" path="res://assets/grid.png" id="3_g5xa3"]
-[ext_resource type="PackedScene" uid="uid://dkqc3of5bdx2i" path="res://src/map/map.tscn" id="4_3k3lb"]
 
-[sub_resource type="ProceduralSkyMaterial" id="ProceduralSkyMaterial_xbryr"]
-sky_horizon_color = Color(0.64625, 0.65575, 0.67075, 1)
-ground_horizon_color = Color(0.64625, 0.65575, 0.67075, 1)
-
-[sub_resource type="Sky" id="Sky_ogrvh"]
-sky_material = SubResource("ProceduralSkyMaterial_xbryr")
-
-[sub_resource type="Environment" id="Environment_ur1u6"]
-background_mode = 2
-sky = SubResource("Sky_ogrvh")
-tonemap_mode = 2
-glow_enabled = true
-
-[sub_resource type="ShaderMaterial" id="ShaderMaterial_lovo8"]
-
-[sub_resource type="PlaneMesh" id="PlaneMesh_x854j"]
-material = SubResource("ShaderMaterial_lovo8")
-size = Vector2(4800, 4800)
-
-[sub_resource type="StandardMaterial3D" id="StandardMaterial3D_fstik"]
-albedo_texture = ExtResource("3_g5xa3")
-uv1_scale = Vector3(300, 300, 300)
-texture_filter = 0
-
-[sub_resource type="StandardMaterial3D" id="StandardMaterial3D_ftw7c"]
-albedo_color = Color(0, 0.654902, 0.658824, 1)
-
-[node name="Main" type="Node3D"]
+[node name="main" type="Node"]
 script = ExtResource("1_2a1rx")
-
-[node name="DirectionalLight3D" type="DirectionalLight3D" parent="."]
-transform = Transform3D(-0.866025, -0.433013, 0.25, 0, 0.5, 0.866025, -0.5, 0.75, -0.433013, 0, 0, 0)
-shadow_enabled = true
-
-[node name="WorldEnvironment" type="WorldEnvironment" parent="."]
-environment = SubResource("Environment_ur1u6")
-
-[node name="Camera3D_FreeCamera" type="Camera3D" parent="."]
-transform = Transform3D(0.707107, -0.5, 0.5, 0, 0.707107, 0.707107, -0.707107, -0.5, 0.5, 20, 10, 20)
-current = true
-script = ExtResource("2_ray12")
-
-[node name="Floor" type="MeshInstance3D" parent="."]
-mesh = SubResource("PlaneMesh_x854j")
-surface_material_override/0 = SubResource("StandardMaterial3D_fstik")
-
-[node name="OriginCube3D" type="CSGBox3D" parent="."]
-transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0.504672, 0)
-material = SubResource("StandardMaterial3D_ftw7c")
-
-[node name="UI" type="Control" parent="."]
-layout_mode = 3
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-grow_horizontal = 2
-grow_vertical = 2
-
-[node name="Map" parent="UI" instance=ExtResource("4_3k3lb")]
-visible = false
-layout_mode = 0
-offset_right = 128.0
-offset_bottom = 128.0
-
-[node name="Label_FPS" type="Label" parent="UI"]
-unique_name_in_owner = true
-layout_mode = 1
-anchors_preset = 1
-anchor_left = 1.0
-anchor_right = 1.0
-offset_left = -157.0
-offset_top = 8.0
-offset_right = -8.0
-offset_bottom = 34.0
-grow_horizontal = 0
-text = "FPS"
-horizontal_alignment = 2
-
-[node name="Label_ParcelPosition" type="Label" parent="UI"]
-unique_name_in_owner = true
-layout_mode = 1
-anchors_preset = 1
-anchor_left = 1.0
-anchor_right = 1.0
-offset_left = -157.0
-offset_top = 32.0
-offset_right = -8.0
-offset_bottom = 58.0
-grow_horizontal = 0
-text = "FPS"
-horizontal_alignment = 2
-
-[node name="Label_Info" type="Label" parent="UI"]
-layout_mode = 1
-anchors_preset = 2
-anchor_top = 1.0
-anchor_bottom = 1.0
-offset_left = 8.0
-offset_top = -34.0
-offset_right = 386.0
-offset_bottom = -8.0
-grow_vertical = 0
-text = "Move with WASD, camera with holding right-click"
-
-[node name="Panel_SpawnScene" type="Panel" parent="UI"]
-layout_mode = 1
-anchors_preset = 3
-anchor_left = 1.0
-anchor_top = 1.0
-anchor_right = 1.0
-anchor_bottom = 1.0
-offset_left = -432.0
-offset_top = -179.0
-offset_right = -208.0
-offset_bottom = -8.0
-grow_horizontal = 0
-grow_vertical = 0
-
-[node name="Button_AddScene" type="Button" parent="UI/Panel_SpawnScene"]
-layout_mode = 1
-anchors_preset = 7
-anchor_left = 0.5
-anchor_top = 1.0
-anchor_right = 0.5
-anchor_bottom = 1.0
-offset_left = -97.0
-offset_top = -39.0
-offset_right = 95.0
-offset_bottom = -8.0
-grow_horizontal = 2
-grow_vertical = 0
-text = "Spawn scene"
-
-[node name="X" type="SpinBox" parent="UI/Panel_SpawnScene"]
-layout_mode = 1
-anchors_preset = 5
-anchor_left = 0.5
-anchor_right = 0.5
-offset_left = -97.5313
-offset_top = 80.0
-offset_right = -14.4688
-offset_bottom = 115.0
-grow_horizontal = 2
-min_value = -10.0
-max_value = 10.0
-
-[node name="Z" type="SpinBox" parent="UI/Panel_SpawnScene"]
-layout_mode = 1
-anchors_preset = 5
-anchor_left = 0.5
-anchor_right = 0.5
-offset_left = 14.4688
-offset_top = 80.0
-offset_right = 97.5313
-offset_bottom = 115.0
-grow_horizontal = 2
-min_value = -10.0
-max_value = 10.0
-
-[node name="Label" type="Label" parent="UI/Panel_SpawnScene"]
-layout_mode = 1
-anchors_preset = 5
-anchor_left = 0.5
-anchor_right = 0.5
-offset_left = -54.5
-offset_top = 48.0
-offset_right = 54.5
-offset_bottom = 74.0
-grow_horizontal = 2
-text = "Parcel coords:"
-
-[node name="OptionButton_SceneToSpawn" type="OptionButton" parent="UI/Panel_SpawnScene"]
-layout_mode = 1
-anchors_preset = 5
-anchor_left = 0.5
-anchor_right = 0.5
-offset_left = -96.0
-offset_top = 8.0
-offset_right = 96.0
-offset_bottom = 39.0
-grow_horizontal = 2
-item_count = 2
-selected = 0
-popup/item_0/text = "parent_list"
-popup/item_0/id = 0
-popup/item_1/text = "cube_wave"
-popup/item_1/id = 1
-
-[node name="Panel_Scenes" type="Panel" parent="UI"]
-layout_mode = 1
-anchors_preset = 3
-anchor_left = 1.0
-anchor_top = 1.0
-anchor_right = 1.0
-anchor_bottom = 1.0
-offset_left = -194.0
-offset_top = -177.0
-offset_right = -8.0
-offset_bottom = -6.0
-grow_horizontal = 0
-grow_vertical = 0
-
-[node name="ItemList_Scenes" type="ItemList" parent="UI/Panel_Scenes"]
-layout_mode = 1
-anchors_preset = 15
-anchor_right = 1.0
-anchor_bottom = 1.0
-offset_left = 16.0
-offset_top = 27.0
-offset_right = -16.0
-offset_bottom = -47.0
-grow_horizontal = 2
-grow_vertical = 2
-
-[node name="Button_DeleteScene" type="Button" parent="UI/Panel_Scenes"]
-layout_mode = 1
-anchors_preset = 12
-anchor_top = 1.0
-anchor_right = 1.0
-anchor_bottom = 1.0
-offset_left = 15.0
-offset_top = -39.0
-offset_right = -17.0
-offset_bottom = -8.0
-grow_horizontal = 2
-grow_vertical = 0
-text = "Delete"
-
-[node name="Label" type="Label" parent="UI/Panel_Scenes"]
-layout_mode = 1
-anchors_preset = 10
-anchor_right = 1.0
-offset_bottom = 26.0
-grow_horizontal = 2
-text = "Scenes"
-horizontal_alignment = 1
-vertical_alignment = 1
-
-[node name="Control" type="Control" parent="UI"]
-anchors_preset = 0
-offset_right = 40.0
-offset_bottom = 40.0
-
-[connection signal="pressed" from="UI/Panel_SpawnScene/Button_AddScene" to="." method="_on_add_button_pressed"]
-[connection signal="pressed" from="UI/Panel_Scenes/Button_DeleteScene" to="." method="_on_button_delete_scene_pressed"]
