+++ godot/src/ui/map/map_data.png
