+++ rust/decentraland-godot-lib/src/scene_runner/components/mod.rs
@@ -1,2 +1,3 @@
+pub mod gltf_container;
 pub mod mesh_renderer;
 pub mod transform_and_parent;
